#89, #2584, opj#2217

+++ ita_root/ita_ag_ansible_execution/setup.sh
@@ -1087,10 +1087,15 @@ install_agent_service(){
         ENV_TMP_PATH="${default_env_values['REFERENCE_ENVPATH']}"
     fi
 
-    if [ -z $ENV_TMP_PATH ] && [ "`realpath ${ENV_TMP_PATH}`" != "`realpath ${ENV_PATH}`" ]; then
-        if [ -f $ENV_PATH ]; then
-            rm -rf $ENV_PATH
+    if [ "${default_env_values['REFERENCE_ENVPATH']}" != "" ]; then
+        if [ -z $ENV_TMP_PATH ] && [ "`realpath ${ENV_TMP_PATH}`" != "`realpath ${ENV_PATH}`" ]; then
+            if [ -f $ENV_PATH ]; then
+                rm -rf $ENV_PATH
+            fi
+            info "cp -rf $ENV_TMP_PATH $ENV_PATH"
+            cp -rf $ENV_TMP_PATH $ENV_PATH
         fi
+    else
         info "cp -rf $ENV_TMP_PATH $ENV_PATH"
         cp -rf $ENV_TMP_PATH $ENV_PATH
     fi
